requirements.txt
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-PyMongo==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
PyJWT==2.10.0
pymongo==4.10.1
python-dotenv==1.0.1
setuptools==75.1.0
Werkzeug==3.1.3
wheel==0.44.0
